requirements.txt
asgiref==3.8.1
django==5.1.1
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==10.4.0
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
